requirements.txt
django==2.2
djangorestframework
django-mssql-backend
django-filter
